Disable failing refcount tests
- sys.getrefcount should only be used to check relative refcounts
- update bindings for integer test

+++ test/integers.wrap.cxx
@@ -36,7 +36,7 @@ static auto const fun_1 = c2py::dispatcher_f_kw_t{c2py::cfun([](short x) { retur
 static auto const fun_2 = c2py::dispatcher_f_kw_t{c2py::cfun([](int x) { return identity_int32(x); }, "x")};
 
 // identity_int64
-static auto const fun_3 = c2py::dispatcher_f_kw_t{c2py::cfun([](long x) { return identity_int64(x); }, "x")};
+static auto const fun_3 = c2py::dispatcher_f_kw_t{c2py::cfun([](long long x) { return identity_int64(x); }, "x")};
 
 // identity_uint16
 static auto const fun_4 = c2py::dispatcher_f_kw_t{c2py::cfun([](unsigned short x) { return identity_uint16(x); }, "x")};
@@ -45,7 +45,7 @@ static auto const fun_4 = c2py::dispatcher_f_kw_t{c2py::cfun([](unsigned short x
 static auto const fun_5 = c2py::dispatcher_f_kw_t{c2py::cfun([](unsigned int x) { return identity_uint32(x); }, "x")};
 
 // identity_uint64
-static auto const fun_6 = c2py::dispatcher_f_kw_t{c2py::cfun([](unsigned long x) { return identity_uint64(x); }, "x")};
+static auto const fun_6 = c2py::dispatcher_f_kw_t{c2py::cfun([](unsigned long long x) { return identity_uint64(x); }, "x")};
 
 static const auto doc_d_0 = fun_0.doc(R"DOC()DOC");
 static const auto doc_d_1 = fun_1.doc(R"DOC()DOC");
@@ -54,7 +54,6 @@ static const auto doc_d_3 = fun_3.doc(R"DOC()DOC");
 static const auto doc_d_4 = fun_4.doc(R"DOC()DOC");
 static const auto doc_d_5 = fun_5.doc(R"DOC()DOC");
 static const auto doc_d_6 = fun_6.doc(R"DOC()DOC");
-
 //--------------------- module function table  -----------------------------
 
 static PyMethodDef module_methods[] = {
